Restructuring of koin
