+++ app/src/main/java/barant/curso/simpsonsapi/core/di/FeatureModule.kt
@@ -1,0 +1,8 @@
+package barant.curso.simpsonsapi.core.di
+
+import barant.curso.simpsonsapi.feature.character.di.characterModule
+import org.koin.dsl.module
+
+val featureModule = module {
+    includes(characterModule)
+}
